Fix BaseClaimForm to create image and debate
