+++ src/components/Claim/CreateClaim/ClaimUploadImage.tsx
@@ -11,6 +11,7 @@ import colors from "../../../styles/colors";
 import ImageUpload, { UploadFile } from "../../ImageUpload";
 import Label from "../../Label";
 import BaseClaimForm from "./BaseClaimForm";
+import { useBaseClaimForm } from "./UseBaseClaimForm";
 
 const ClaimUploadImage = () => {
     const { t } = useTranslation();
@@ -18,48 +19,55 @@ const ClaimUploadImage = () => {
     const [, send] = useAtom(createClaimMachineAtom);
 
     const [fileList, setFileList] = useState<UploadFile[]>([]);
-    const [imageError, setImageError] = useState(false);
-    const [isLoading, setIsLoading] = useState(false);
+    const {
+        title, setTitle, date, setDate, sources, setSources, recaptcha, setRecaptcha, isLoading, setIsLoading, errors, setErrors, imageError, setImageError, clearError, validateFields
+    } = useBaseClaimForm({ shouldValidateContent: false });
 
     const handleSubmit = (values: any) => {
-        if (fileList.length > 0) {
-            setIsLoading(true);
-            const formData = new FormData();
+        const newErrors = validateFields();
 
-            fileList.forEach((file) => {
-                if (file.originFileObj) {
-                    formData.append("files", file.originFileObj);
+        if (Object.keys(newErrors).length > 0 || fileList.length === 0) {
+            if (fileList.length === 0) {
+                setImageError(true);
+            }
+            setErrors(newErrors);
+            return;
+        }
+
+        setIsLoading(true);
+        const formData = new FormData();
+
+        fileList.forEach((file) => {
+            if (file.originFileObj) {
+                formData.append("files", file.originFileObj);
+            }
+        });
+
+        ImageApi.uploadImage(formData, t)
+            .then((imagesUploaded) => {
+                setImageError(false);
+
+                const claimData = {
+                    ...values,
+                    content: imagesUploaded[0],
+                };
+
+                send({
+                    type: CreateClaimEvents.persist,
+                    claimData,
+                    t,
+                    router,
+                });
+            })
+            .catch((err) => {
+                setIsLoading(false);
+                if (err.response?.status === 303) {
+                    const seeOtherTarget = err.response.data.target;
+                    if (seeOtherTarget) {
+                        router.push(seeOtherTarget);
+                    }
                 }
             });
-
-            ImageApi.uploadImage(formData, t)
-                .then((imagesUploaded) => {
-                    setImageError(false);
-
-                    const claimData = {
-                        ...values,
-                        content: imagesUploaded[0],
-                    };
-
-                    send({
-                        type: CreateClaimEvents.persist,
-                        claimData,
-                        t,
-                        router,
-                    });
-                })
-                .catch((err) => {
-                    setIsLoading(false);
-                    if (err.response?.status === 303) {
-                        const seeOtherTarget = err.response.data.target;
-                        if (seeOtherTarget) {
-                            router.push(seeOtherTarget);
-                        }
-                    }
-                });
-        } else {
-            setImageError(true);
-        }
     };
 
     const handleFileChange = (newFileList: UploadFile[]) => {
@@ -100,6 +108,16 @@ const ClaimUploadImage = () => {
                 handleSubmit={handleSubmit}
                 isLoading={isLoading}
                 dateExtraText={t("claimForm:dateFieldHelpImage")}
+                errors={errors}
+                clearError={clearError}
+                recaptcha={recaptcha}
+                setRecaptcha={setRecaptcha}
+                setTitle={setTitle}
+                date={date}
+                setDate={setDate}
+                setSources={setSources}
+                title={title}
+                sources={sources}
                 content={
                     <Grid container style={{ marginBottom: 24 }}>
                         <Label required>{t("claimForm:fileInputButton")}</Label>
